Extract data generation into BearGenerator
As it turns out, generating data becomes more complex once you introduce
computed values (such as a block size) - which depends on the content of
other values. Dealing with that requires to keep track of the current
variable, all other variables, and all the values from the current
variability vector.

Extracting that into a new class not only reduces the size of bear.c
and overall complexity.